requirements.txt
prometheus-client==0.19.0
psycopg2-binary==2.9.9
apache-airflow==2.7.0
apache-airflow-providers-docker==3.7.0